General Enemy/Weapon Fixes
All bullets should now inherit damage, speed and lifetime values properly.
Enemies will no longer spawn in walls.

+++ Corpses/Assets/Scripts/WeaponScripts/LegacyScripts/StaffProjectileScript.cs
@@ -8,7 +8,7 @@ using UnityEngine;
 public class StaffProjectileScript : MonoBehaviour
 {
     Rigidbody2D rb;
-    public int bulletLifetime;
+    public float bulletLifetime;
 
     private Vector2 lastVelocity;
     private float currentSpeed;
